Updates to productSlice, Product component, index.tsx, & types.ts
