+++ src/client/product/Products.tsx
@@ -1,0 +1,7 @@
+// Products Component for managing products
+
+
+export function Products() {
+
+}
+
