+++ src/types.ts
@@ -2,6 +2,14 @@ export type params =  string | number
 import { Request, Response, NextFunction } from 'express';
 
 
+// Represents a product & its details/properties
+export interface Product {
+    id: number;  // unique identifier for each product
+    product_name: string;
+    price: number;
+    sold_out: boolean;  // availability.  True if sold out.
+}
+
 export interface menu {
     getMenuItems: (req: Request, res: Response, next: NextFunction) => Promise<void>;
     addMenuItem: (req: Request, res: Response, next: NextFunction) => Promise<void>;
